Fixed API creation logic
As reported on the ML, the "create API" form was not actually submitting. This
was due to a change in the template which prevented the modal from having the
same scope as the API list in the navbar.

This patch separates the creation vs list logic into separate controllers, and
has the creation logic emit a signal that the list logic intercepts so that the
navbar will dynamically update on creation.