Better asset src/dist separation
Instead of having src and dist both served publically, one or the other should
be available only. This commit separates the two, and introduces a tool for
switching between them.

Additionally, this separation should correct issues with the base path when the
app is being run from a subpath of the webroot.