Fixes for API list stickiness
Returning to the API list page caused the versions to no longer be present in
the links. On inspection, it was never clear why they were _ever_ present
properly.

Made some fixes to the HAL service to ensure that stripping links and embedded
resources in arrays of resources works properly, which was one source of
problems.

Removed the "ApiListController", as it isn't currently used.

Updated the "ApiController" to no longer accept the apis list, and instead fetch
it on first load; subsequent calls will use cached data, unless something
triggers a list update. The "default_version" property of each API is used to
create the links if no versions are present in the API.